Added Class, Fixed up Driver, Cleaned up View Class
Added RateSubjectCourse that deals with subject and course select based
on user input and returns to driver. The driver wasn't doing much, but
I now have a good idea for what its purpose will be. It will control
input from the view and send it through the subjectCourse class and
then return to the view to display. The view class used to do a lot on
its own, but now it calls the driver and so is more efficient.