No Blank Screens!
I changed the pickedSubject variable from a local to an instance
variable which allows me to return null if what the user typed in is
not a subject or course. Before when it was null, there would be a
blank screen, but now if it returns null, it defaults to the instance
variable as it was before the user typed anything in. Effectively, this
means that if null is returned, the screen stays the same.