Added Back Class, Survey Questions, 2D Arrays
Added new class called RateSaveSurvey. It creates the 2D arrays and
subsequent arrays to fill it. It then uses the users subject and course
choice to iterate through the subject's array to find the index of the
right course. Using that index, I will next work on saving the user's
response to the survey in that space in the 2D array.