Survey Almost Displays Results!
I keep getting some errors after the user has completed the survey. I'm
not sure if the error has to do with adding user data to the ArrayList
or calling the ArrayLists. If I fix that, then theoretically the data
for the course the user chose should appear (question averages).